use serverを使用するとエラーが出るのでapiを使用する方法を採用しました
