+++ src/app/admin/page.tsx
@@ -1,14 +1,14 @@
 "use client";
 
-import { StartGame } from "@/logics/GameStartFlagHandler";
-export default function Admin() {
+import { handleStartGame } from "@/logics/GameStartFlagHandler";
 
+export default function Admin() {
   return (
     <div>
       <div className="text-center">
         <div>管理画面</div>
         <button
-          onClick={StartGame}>
+          onClick={handleStartGame}>
           ゲームを開始
         </button>
       </div>
